pyproject.toml
[project]
name = "soccer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Matt Yorkilous", email = "matt_yorkilous@yahoo.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "kagglehub>=0.4.0",
    "numpy>=2.4.1",
    "polars-lts-cpu>=1.33.1",
    "scikit-learn>=1.8.0",
]

[project.scripts]
soccer = "soccer:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
